Conflicts:
	src/corelib/global/minimum-linux.S
	src/network/access/qhttpthreaddelegate.cpp
	src/widgets/kernel/qwidgetwindow.cpp

+++ src/gui/math3d/qquaternion.h
@@ -210,8 +210,8 @@ inline QQuaternion QQuaternion::inverted() const
                  double(yp) * double(yp) +
                  double(zp) * double(zp);
     if (!qFuzzyIsNull(len))
-        return QQuaternion(double(wp) / len, double(-xp) / len,
-                           double(-yp) / len, double(-zp) / len);
+        return QQuaternion(float(double(wp) / len), float(double(-xp) / len),
+                           float(double(-yp) / len), float(double(-zp) / len));
     return QQuaternion(0.0f, 0.0f, 0.0f, 0.0f);
 }
